:bug: MainViewModel: Ensure lifecycle doesn't break
This class is not an extension of [ViewModel] because the [ViewModel]s lifecycle is broken with launchers (doing the home gesture the first time recreates the view model). Since our [MainActivity] lives as long as the launcher is running if the launcher is set as default, it's okay that our [MainViewModel] lives as long as the launcher is running.